Refactor and migrate Footy, Forecast, and Reload commands
Migrated Footy, Forecast, and Reload commands from OldCommands to new class-based structure. Updated BetaReload pattern to match '/^reload$/'. Moved getLocale utility from OldCommands/Utils.php to BotUtils.php for shared access. Removed legacy command files and updated command handler accordingly.
