--- Commands/OldCommands/Reload.php
@@ -1,11 +1,0 @@
-<?php
-
-	function Reload($message) { 
-	
-		if (isAdmin($message->author->id) && getenv('BETA') !== 'true') {
-			die();
-		}
-		
-	}
-	
-?>
